ds18b20: hackishly detect if the conversion is valid
Before a temperature conversion command is issued to the device, the
temperature reading is 85C. In the case where the device loses power
between issuing the temperature conversion command and reading the
temperature, an invalid result (85C) is returned. Even the checksum
matches. Before sending the temperature conversion command, this code
writes data to the th and tl registers that can be verified when the
temperature is read. If the byte is not the same as the one that was
written, the device power cycled and the temperature result is invalid.

Definitely hackish, but seems reliable. This would probably not be
necessary if my cable were not somewhat loose :o.